+++ containers/Home/Skill/styles.tsx
@@ -1,9 +1,14 @@
 import styled from 'styled-components'
-import { BackGroundColor, FontColor, GridBreakpoints } from 'style/Theme'
+import {
+    BackGroundColor,
+    FontColor,
+    GridBreakpoints,
+    ResponsiveBetween,
+} from 'style/Theme'
 import { ProcessBar } from './ProcessBar'
 
 export const StyledSkill = styled.div`
-    @media (min-width: ${GridBreakpoints.sm}) and (max-width: ${GridBreakpoints.lg}) {
+    ${ResponsiveBetween('sm', 'lg')} {
         width: 50%;
         padding: 10px;
     }
