+++ containers/Home/index.ts
@@ -1,9 +1,1 @@
-export { About } from './About'
-export { Contact } from './Contact'
-export { Language } from './Language'
-export { Lib } from './Lib'
-export { Skill } from './Skill'
-export { Achievement } from './Achievement'
-export { Service } from './Service'
 export { HomePage } from './Home'
-export { Education } from './Education'
