+++ hooks/index.ts
@@ -1,1 +1,2 @@
 export { useIsBrowser } from './useIsBrowser'
+export { useIsPrintMode } from './useIsPrintMode'
