+++ pages/index.tsx
@@ -1,4 +1,4 @@
-import { HomePage } from '@containers/Home/Home'
+import { HomePage } from '@containers/Home'
 
 const Home: React.FC = () => {
     return <HomePage />
